Time: 305 ms (64.29%), Space: 114.2 MB (7.14%) - LeetHub

+++ 2501-longest-square-streak-in-an-array/2501-longest-square-streak-in-an-array.cpp
@@ -5,7 +5,7 @@ public:
       map<int, int> mp;
       for(auto &i: v) mp[i]++;
       vector<pair<int, int>> ans(n);
-      int maxele = *max_element(v.begin(), v.end());
+      double maxele = *max_element(v.begin(), v.end());
       maxele = sqrt(maxele);
       for(auto &p: mp){
         ans[idx].first = p.first;
